Fix signed left-shift UB in bitmask variables and hoist constant masks
- Change all profMask/fullProfMask variables from int to uint32_t and
  use uint32_t(1) literal in shifts across SimThread.cpp,
  ProductionSystem.cpp, and RandomEventSystem.cpp to eliminate
  signed left-shift undefined behavior.
- Hoist fullProfMask computation above per-settlement lambdas in
  SimThread.cpp and RandomEventSystem.cpp since it depends only on
  the schema and is constant across iterations.
- Move the assert(professions.size() < 32) from AgentDecisionSystem
  (checked per-tick) into WorldSchema::BuildMaps() (checked once at
  load time).

+++ src/ECS/Systems/ProductionSystem.cpp
@@ -143,7 +143,7 @@ void ProductionSystem::Update(entt::registry& registry, float realDt, const Worl
             if (const auto* prof = registry.try_get<Profession>(e)) {
                 if (prof->type >= 0 && prof->type < (int)schema.professions.size()
                     && schema.professions[prof->type].producesResource != INVALID_ID)
-                    profDiversity[hs.settlement] |= (1 << prof->type);
+                    profDiversity[hs.settlement] |= (uint32_t(1) << prof->type);
             }
             workers[hs.settlement] += workerContrib;
             workerHeadCount[hs.settlement]++;
@@ -157,9 +157,9 @@ void ProductionSystem::Update(entt::registry& registry, float realDt, const Worl
 
     // ---- Settlement profession diversity bonus ----
     // Settlements with all producing professions get +3% production.
-    int fullProfMask = 0;
+    uint32_t fullProfMask = 0;
     for (auto& pd : schema.professions)
-        if (pd.producesResource != INVALID_ID) fullProfMask |= (1 << pd.id);
+        if (pd.producesResource != INVALID_ID) fullProfMask |= (uint32_t(1) << pd.id);
     {
         static std::map<entt::entity, int> s_diverseLogged;
         for (auto& [settl, w] : workers) {
